requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
pdfplumber>=0.7.0
openpyxl>=3.1.0

# AI and ML
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
google-generativeai>=0.3.0
faiss-cpu>=1.7.4

# Visualization
plotly>=5.15.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
